Add debounce option for fetching search results
This is a minor addition I made that allows for a manual debounce before sending a query, while preserving the default value.

The code is identical to one I did for a minified LocationIQ geocoder. Please note that I have not tested it in maplibre, but the code is incredibly simple, and I am offering it here to be polished if needed, since it seems like a no-brainer feature.

I have no tests to offer, and wouldn't know how to write them. I hope this feature will be considered for use.

Thank you.